Fixed Obtaining Peer Information

+++ x42Client/RestClient/Responses/GetPeerInfoResponse.cs
@@ -7,23 +7,23 @@
         public string addrlocal { get; set; }
         public string services { get; set; }
         public bool relaytxes { get; set; }
-        public int lastsend { get; set; }
+        public uint lastsend { get; set; }
         public int lastrecv { get; set; }
-        public int bytessent { get; set; }
-        public int bytesrecv { get; set; }
-        public int conntime { get; set; }
-        public int timeoffset { get; set; }
-        public int pingtime { get; set; }
-        public int minping { get; set; }
-        public int pingwait { get; set; }
+        public uint bytessent { get; set; }
+        public uint bytesrecv { get; set; }
+        public uint conntime { get; set; }
+        public uint timeoffset { get; set; }
+        public uint pingtime { get; set; }
+        public uint minping { get; set; }
+        public uint pingwait { get; set; }
         public int version { get; set; }
         public string subver { get; set; }
         public bool inbound { get; set; }
         public bool addnode { get; set; }
-        public int startingheight { get; set; }
+        public ulong startingheight { get; set; }
         public int banscore { get; set; }
-        public int synced_headers { get; set; }
-        public int synced_blocks { get; set; }
+        public ulong synced_headers { get; set; }
+        public ulong synced_blocks { get; set; }
         public bool whitelisted { get; set; }
         public object inflight { get; set; }
         public object bytessent_per_msg { get; set; }
